fix java test in CI
